+++ Changelog.md
@@ -1,6 +1,8 @@
 # 2.0 Series
 ## 2.0.0 beta 6
- * add option to hide mouseover dots for line chart, by Davis Ford ([#735](https://github.com/dc-js-dc.js-issues/735))
+ * heatmap box titles were not updated on redraw, by hhravn ([#798](https://github.com/dc-js/dc.js/pull/798))
+ * apply ordering to ordinal bar chart, by Mihai Hodorogea ([#766](https://github.com/dc-js/dc.js/pull/766) / [#772](https://github.com/dc-js/dc.js/issues/772))
+ * add option to hide mouseover dots for line chart, by Davis Ford ([#735](https://github.com/dc-js/dc.js/pull/735))
  * adding example bar-extra-line, overlaying a line with a renderlet
 
 ## 2.0.0 beta 5